+++ src/OpenSimCreator/Widgets/BasicWidgets.hpp
@@ -36,7 +36,7 @@ namespace osc
     void DrawSelectOwnerMenu(VirtualModelStatePair&, OpenSim::Component const&);
     bool DrawWatchOutputMenu(MainUIStateAPI&, OpenSim::Component const&);
     void DrawSimulationParams(ParamBlock const&);
-    void DrawSearchBar(std::string&, int maxLen);
+    void DrawSearchBar(std::string&);
     void DrawOutputNameColumn(
         VirtualOutputExtractor const& output,
         bool centered = true,